Move CDP files from src/bidiMapper to src/
