+++ src/cdp/index.ts
@@ -1,0 +1,1 @@
+export { CdpClient, connectCdp } from './cdpClient';
