+++ src/tsconfig.json
@@ -1,8 +1,6 @@
 {
   "extends": "./configs/tsconfig.common.json",
   "compilerOptions": {
-    "outDir": "./.build",
-    "module": "ESNext"
-  },
-  "exclude": ["bidiMapper"]
+    "outDir": "./.build"
+  }
 }
